+++ api/fe.js
@@ -1,0 +1,11 @@
+const { generateScript } = require('../forward_engineering/generateScript');
+const { generateViewScript } = require('../forward_engineering/generateViewScript');
+const { generateContainerScript } = require('../forward_engineering/generateContainerScript');
+const { isDropInStatements } = require('../forward_engineering/isDropInStatements');
+
+module.exports = {
+	generateScript,
+	generateViewScript,
+	generateContainerScript,
+	isDropInStatements,
+};
