+++ forward_engineering/generateScript.js
@@ -1,0 +1,61 @@
+const _ = require('lodash');
+const sqlFormatter = require('sql-formatter');
+const { getAlterScript } = require('./helpers/alterScriptFromDeltaHelper');
+const { getDatabaseStatement } = require('./helpers/databaseHelper');
+const { getTableStatement } = require('./helpers/tableHelper');
+const { getIndexes } = require('./helpers/indexHelper');
+const { buildScript } = require('./helpers/buildScript');
+
+const generateScript = (data, logger, callback, app) => {
+	try {
+		const jsonSchema = JSON.parse(data.jsonSchema);
+		const modelDefinitions = JSON.parse(data.modelDefinitions);
+		const internalDefinitions = JSON.parse(data.internalDefinitions);
+		const externalDefinitions = JSON.parse(data.externalDefinitions);
+		const containerData = data.containerData;
+		const entityData = data.entityData;
+		const areColumnConstraintsAvailable = data.modelData[0].dbVersion.startsWith('3');
+		const areForeignPrimaryKeyConstraintsAvailable = !data.modelData[0].dbVersion.startsWith('1');
+		const needMinify = _.get(data, 'options.additionalOptions', []).find(option => option.id === 'minify')?.value;
+
+		if (data.isUpdateScript) {
+			const definitions = [modelDefinitions, internalDefinitions, externalDefinitions];
+			const scripts = getAlterScript(jsonSchema, definitions, data, app, needMinify, sqlFormatter);
+			callback(null, scripts);
+			return;
+		}
+
+		callback(
+			null,
+			buildScript(needMinify)(
+				getDatabaseStatement(containerData),
+				getTableStatement(
+					containerData,
+					entityData,
+					jsonSchema,
+					[modelDefinitions, internalDefinitions, externalDefinitions],
+					null,
+					areColumnConstraintsAvailable,
+					areForeignPrimaryKeyConstraintsAvailable,
+				),
+				getIndexes(
+					containerData,
+					entityData,
+					jsonSchema,
+					[modelDefinitions, internalDefinitions, externalDefinitions],
+					areColumnConstraintsAvailable,
+				),
+			),
+		);
+	} catch (e) {
+		logger.log('error', { message: e.message, stack: e.stack }, 'Hive Forward-Engineering Error');
+
+		setTimeout(() => {
+			callback({ message: e.message, stack: e.stack });
+		}, 150);
+	}
+};
+
+module.exports = {
+	generateScript,
+};
